backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
openai-whisper==20231117
torch==2.2.0
numpy==1.26.2
transformers==4.36.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
psycopg2-binary==2.9.9
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
ffmpeg-python==0.2.0
librosa>=0.10.0
vosk>=0.3.45
sentencepiece>=0.1.99
protobuf>=4.24.4
pytest>=7.4.3
pytest-asyncio>=0.21.0
httpx>=0.25.2
pytest-cov>=4.1.0
aiofiles==23.2.1
bcrypt==4.0.1
google-auth==2.27.0 